haha

+++ src/main/java/com/example/HotelBookingSystem/service/PaymentServiceImpl.java
@@ -5,6 +5,6 @@ import com.example.HotelBookingSystem.model.Payment;
 import java.math.BigDecimal;
 
 public interface PaymentServiceImpl {
-    Payment createPayment(Long bookingId, BigDecimal amount, String method);
+    Payment createPayment(Integer bookingId, BigDecimal amount, String method);
     Payment updatePayment(String transactionRef, String status);
 }
